requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
PyJWT==2.10.1
pymongo==4.15.3
python-dotenv==1.1.1
pywin32==311; platform_system == "Windows"
Werkzeug==3.1.3
